Move implementations to header files
